update to gold cap work-around

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -2334,7 +2334,7 @@ void ObjectMgr::LoadItemTemplates()
         for (uint8 i = 0; i < itemTemplate.StatsCount; ++i)
         {
             itemTemplate.ItemStat[i].ItemStatType  = uint32(fields[28 + i*2].GetUInt8());
-            itemTemplate.ItemStat[i].ItemStatValue = int32(fields[29 + i*2].GetInt16());
+            itemTemplate.ItemStat[i].ItemStatValue = int32(fields[29 + i*2].GetInt32());
         }
 
         itemTemplate.ScalingStatDistribution = uint32(fields[48].GetUInt16());
@@ -2342,12 +2342,12 @@ void ObjectMgr::LoadItemTemplates()
 
         for (uint8 i = 0; i < MAX_ITEM_PROTO_DAMAGES; ++i)
         {
-            itemTemplate.Damage[i].DamageMin  = fields[50 + i*3].GetFloat();
-            itemTemplate.Damage[i].DamageMax  = fields[51 + i*3].GetFloat();
+            itemTemplate.Damage[i].DamageMin  = fields[50 + i*3].GetUInt32();
+            itemTemplate.Damage[i].DamageMax  = fields[51 + i*3].GetUInt32();
             itemTemplate.Damage[i].DamageType = uint32(fields[52 + i*3].GetUInt8());
         }
 
-        itemTemplate.Armor          = uint32(fields[56].GetUInt16());
+        itemTemplate.Armor          = uint32(fields[56].GetUInt32());
         itemTemplate.HolyRes        = uint32(fields[57].GetUInt8());
         itemTemplate.FireRes        = uint32(fields[58].GetUInt8());
         itemTemplate.NatureRes      = uint32(fields[59].GetUInt8());
